Clean up prior named directories plus neglected Reference is included
